improve md perf + share code
