+++ js/markdown/static/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./StaticMarkdown.svelte";
+export { default as MarkdownCode } from "./MarkdownCode.svelte";
